requirements.txt
aioredis==1.3.1
altgraph==0.17
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.14
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.5
constantly==15.1.0
cors==1.0.1
cryptography==38.0.1
daphne==3.0.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.2.0
email-validator==1.1.3
et-xmlfile==1.1.0
filelock==3.8.0
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
geojson==2.5.0
gevent==21.12.0
greenlet==1.1.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.4
openpyxl==3.0.9
pefile==2019.4.18
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pyOpenSSL==22.0.0
pyowm==3.2.0
PySocks==1.7.1
pytz==2022.2.1
pywin32-ctypes==0.2.0
requests==2.28.1
requests-file==1.5.1
service-identity==21.1.0
six==1.16.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
tldextract==3.3.1
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
Werkzeug==2.0.2
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
